View Bill
Note: when added you will need to add another button from the home screen to use this.

Bugs: nothing is displayed when the view bill pannel is launched
Cause: The rs.get____ commands wont pull the information from the bills table in our database. because of this the program doesn't go past the  rs.get____ command and it just waits there so the table never gets built

I've checked to make sure it wasen't the j table causing the issue by copying lines 72-75 
to line 55. and the result is the screen shot pasted along with the code. The table is displayed correctly without the information from the database.